requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.0
billiard==4.2.4
CacheControl==0.14.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
Django==6.0.1
django-browser-reload==1.21.0
django-ckeditor==6.7.3
django-environ==0.12.0
django-js-asset==3.1.2
django-tailwind==4.4.2
django_celery_results==2.6.0
firebase_admin==7.1.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==24.1.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.4.0
